+++ src/core/InputHandler.java
@@ -1,12 +1,17 @@
 package core;
 
-import java.awt.event.KeyAdapter;
-import java.awt.event.KeyEvent;
+import java.awt.event.*;
 
-public class InputHandler extends KeyAdapter {
+public class InputHandler extends KeyAdapter implements MouseMotionListener, MouseListener {
 
     private boolean leftPressed = false;
     private boolean rightPressed = false;
+
+    private int mouseX = 0;
+    private int mouseY = 0;
+    private boolean mousePressed = false;
+
+    // ===== KEY =====
 
     @Override
     public void keyPressed(KeyEvent e) {
@@ -36,5 +41,50 @@ public class InputHandler extends KeyAdapter {
 
     public boolean isRightPressed() {
         return rightPressed;
+    }
+
+    // ===== MOUSE =====
+    @Override
+    public void mouseMoved(MouseEvent e) {
+        mouseX = e.getX();
+        mouseY = e.getY();
+    }
+
+    @Override
+    public void mouseDragged(MouseEvent e) {
+        mouseX = e.getX();
+        mouseY = e.getY();
+    }
+
+    @Override
+    public void mousePressed(MouseEvent e) {
+        mousePressed = true;
+    }
+
+    @Override
+    public void mouseReleased(MouseEvent e) {
+        mousePressed = false;
+    }
+
+    @Override
+    public void mouseClicked(MouseEvent e) { }
+
+    @Override
+    public void mouseEntered(MouseEvent e) { }
+
+    @Override
+    public void mouseExited(MouseEvent e) { }
+
+    // ===== GETTERS =====
+    public int getMouseX() {
+        return mouseX;
+    }
+
+    public int getMouseY() {
+        return mouseY;
+    }
+
+    public boolean isMousePressed() {
+        return mousePressed;
     }
 }